chore: rename cmd to the tool name
