+++ cmd/bittwister/serve.go
@@ -1,4 +1,4 @@
-package cmd
+package bittwister
 
 import (
 	"fmt"
